fixed error when generating new maze when still generating and added comments
Whenever you pressed the button to generate a new maze while it was still generating before it would give an error that one of the deleted cells could not be accessed anymore. This was not a problem that caused any issues for the code, but with it still needed to be fixed.
Also added comments where they were missing to give necessary clarification about the code